fixing AdminEntryCost input to parseFloat instead of parseInt to allow for less than 1 ETH value on updated costs, adding styling on EntryCards, changing word chocie within text within index page

+++ frontend/components/AdminEntryCost.tsx
@@ -36,7 +36,7 @@ const AdminEntryCost = () => {
       <Input
         type='number'
         value={entryCostValue}
-        onChange={(e) => setEntryCostValue(parseInt(e.target.value))}
+        onChange={(e) => setEntryCostValue(parseFloat(e.target.value))}
       />
       <Web3Button
         contractAddress={RAFFLE_CONTRACT_ADDRESS}
